+++ src/main/java/com/sw1tech/orcamento/Contratos/Repositorios/IRepositorioParceiro.java
@@ -3,6 +3,6 @@ package com.sw1tech.orcamento.Contratos.Repositorios;
 import com.sw1tech.orcamento.Entidades.Parceiro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IRepositorioParceiro extends JpaRepository<Parceiro, Integer> {
+public interface IRepositorioParceiro extends JpaRepository<Parceiro, Long> {
 
 }
