+++ src/main/java/com/sw1tech/orcamento/Contratos/Repositorios/IRepositorioSexo.java
@@ -3,6 +3,6 @@ package com.sw1tech.orcamento.Contratos.Repositorios;
 import com.sw1tech.orcamento.Entidades.Sexo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IRepositorioSexo extends JpaRepository<Sexo, Integer> {
+public interface IRepositorioSexo extends JpaRepository<Sexo, Long> {
 
 }
