+++ src/main/java/com/sw1tech/orcamento/Contratos/Repositorios/IRepositorioUf.java
@@ -3,6 +3,6 @@ package com.sw1tech.orcamento.Contratos.Repositorios;
 import com.sw1tech.orcamento.Entidades.Uf;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IRepositorioUf extends JpaRepository<Uf, Integer> {
+public interface IRepositorioUf extends JpaRepository<Uf, Long> {
 
 }
